Fix HOCR
The Tesseract 3.05+ output contains attributes separated by semicolons for ocr_line and ocrx_word classes. The defect occurs when using the Page Viewer with HOCR and the coordinates of the polygon have been parsed incorrectly, thus not paints the bounding box of lines and words. In this case verifies the existence of more attributes and parsed the coordinates that should appear in the first bbox attribute.